Upgrade to bouncycastle 1.59. Created Maven project.
