+++ src/main/java/ro/mobilPay/payment/request/Bitcoin.java
